+++ landing/content/_meta.js
@@ -2,9 +2,7 @@ export default {
   index: 'Introduction',
   'getting-started': 'Getting Started',
   examples: 'Examples',
-  evm: 'EVM chains',
-  solana: 'Solana',
-  starknet: 'Starknet',
+  chains: 'Chains',
   contributing: 'Contributing',
   changelog: 'Changelog'
 }